Added apperance section template and refactored files
