+++ src/components/profile/sections/AppearanceSection.tsx
@@ -1,0 +1,7 @@
+interface IProps {}
+
+const AppearanceSection = ({}: IProps) => {
+  return <div>AppearanceSection</div>;
+};
+
+export default AppearanceSection;
