+++ src/components/profile/sections/SecuritySection.tsx
@@ -1,7 +1,7 @@
 import { Box, Button, Grid, Typography } from "@mui/material";
-import LogoutAllButton from "../auth/LogoutAllButton";
+import LogoutAllButton from "../../auth/LogoutAllButton";
 import KeyIcon from "@mui/icons-material/Key";
-import DeleteButton from "../auth/DeleteButton";
+import DeleteButton from "../../auth/DeleteButton";
 
 interface IProps {}
